Fix Camera spot calculation; Fix auto problems; Fix width\height order

+++ include/ray-tracing/render/renderer.hpp
@@ -18,7 +18,7 @@ class Renderer {
   Frame RenderScene(Scene* scene, const Resolution& resolution);
 
  private:
-  Color RenderPixel(size_t x, size_t y) const;
+  Color RenderPixel(double x, double y) const;
   Collision FindClosestCollision(const Ray& ray) const;
   double GetBrightness(Vector position) const;
   static double CalculateBrightness(double cosinus, double distance);
